requirements.txt
langchain
python-dotenv
streamlit
tiktoken
faiss-cpu
InstructorEmbedding